Cập nhật lại cấu trúc code, Sửa lại Update Product
Cập nhật cấu trúc của AuthService, MailService, thêm DeviceService

+++ Source/FN.Application/Catalog/Product/ProductManageService.cs
@@ -15,6 +15,7 @@ namespace FN.Application.Catalog.Product
 {
     public class ProductManageService : IProductManageService
     {
+        private const string ROOT = "product";
         private readonly AppDbContext _db;
         private readonly IMapper _mapper;
         private readonly IImageService _image;
@@ -104,7 +105,10 @@ namespace FN.Application.Catalog.Product
                     .Where(x => x.Id == product.Id)
                     .FirstOrDefaultAsync();
 
-                _db.Items.Remove(item);
+                if (item != null)
+                {
+                    _db.Items.Remove(item);
+                }
 
                 await _db.SaveChangesAsync();
 
@@ -118,6 +122,7 @@ namespace FN.Application.Catalog.Product
             }
         }
 
+
         public async Task<ApiResult<bool>> Delete(int itemId, int userId)
         {
             var item = await _db.Items.FirstOrDefaultAsync(x => x.Id == itemId && x.UserId == userId);
@@ -129,5 +134,100 @@ namespace FN.Application.Catalog.Product
             await _db.SaveChangesAsync();
             return new ApiSuccessResult<bool>();
         }
+        public async Task<ApiResult<bool>> DeleteImage(DeleteProductImagesRequest request)
+        {
+            using var transaction = await _db.Database.BeginTransactionAsync();
+            try
+            {
+                var imagesToDelete = await _db.ProductImages.Include(x => x.ProductDetail)
+                    .ThenInclude(x => x.Item)
+                    .Where(pi => request.ImageIds.Contains(pi.Id)).ToListAsync();
+                if (imagesToDelete == null || !imagesToDelete.Any()) return new ApiErrorResult<bool>("Không tìm thấy ảnh cần xóa.");
+
+                foreach (var image in imagesToDelete)
+                {
+                    await _image.DeleteImageInFolder(image.PublicId, Folder(image.ProductDetail.Item.Id));
+                    _db.ProductImages.Remove(image);
+                }
+
+                await _db.SaveChangesAsync();
+                await transaction.CommitAsync();
+                return new ApiSuccessResult<bool>(true);
+            }
+            catch (Exception ex)
+            {
+                await transaction.RollbackAsync();
+                return new ApiErrorResult<bool>("Xóa ảnh thất bại: " + ex.Message);
+            }
+        }
+        public async Task<ApiResult<bool>> UpdateCombined(CombinedUpdateRequest request, int itemId, int productId, int userId)
+        {
+            // Cập nhật Item
+            var item = await _db.Items.FirstOrDefaultAsync(x => x.Id == itemId && x.UserId == userId);
+            if (item == null) return new ApiErrorResult<bool>("Không tìm thấy sản phẩm");
+            if (!string.IsNullOrEmpty(request.Title))
+            {
+                string code = StringHelper.GenerateProductCode(request.Title);
+                item.Title = request.Title;
+                item.Code = code;
+                item.SeoAlias = StringHelper.GenerateSeoAlias(request.Title);
+                item.SeoTitle = request.Title;
+                item.NormalizedTitle = StringHelper.NormalizeString(request.Title);
+            }
+            if (!string.IsNullOrEmpty(request.Description))
+                item.Description = request.Description;
+            if (!string.IsNullOrEmpty(request.Keywords))
+                item.Keywords = request.Keywords;
+            if (request.Thumbnail != null)
+            {
+                string? newThumbnail = await _image.UploadImage(request.Thumbnail, item.Code, Folder(item.Id.ToString()));
+                if (!string.IsNullOrEmpty(newThumbnail)) item.Thumbnail = newThumbnail;
+            }
+            item.ModifiedDate = DateTime.Now;
+            _db.Items.Update(item);
+
+            // Cập nhật ProductDetail
+            var product = await _db.ProductDetails.FirstOrDefaultAsync(x => x.ItemId == itemId && x.Id == productId);
+            if (product == null) return new ApiErrorResult<bool>("Không tìm thấy chi tiết sản phẩm");
+            product.Status = request.Status;
+            if (!string.IsNullOrEmpty(request.Detail))
+                product.Detail = request.Detail;
+            if (!string.IsNullOrEmpty(request.Note))
+                product.Note = request.Note;
+            if (!string.IsNullOrEmpty(request.Version))
+                product.Version = request.Version;
+            product.CategoryId = request.CategoryId;
+            _db.ProductDetails.Update(product);
+
+            if (request.NewImages != null)
+            {
+                foreach (var file in request.NewImages)
+                {
+                    var publicId = _image.GenerateId();
+                    var resultUpload = await _image.UploadImage(file, publicId, Folder(item.Id.ToString()));
+                    if (!string.IsNullOrEmpty(resultUpload))
+                    {
+                        var newImage = new ProductImage
+                        {
+                            Caption = file.FileName,
+                            ImageUrl = resultUpload,
+                            PublicId = publicId,
+                            ProductDetailId = product.Id
+                        };
+                        _db.ProductImages.Add(newImage);
+                    }
+
+                }
+            }
+            // Lưu thay đổi vào cơ sở dữ liệu
+            await _db.SaveChangesAsync();
+
+            return new ApiSuccessResult<bool>(true);
+        }
+        string Folder(string code)
+        {
+            return $"{ROOT}/{code}";
+        }
+
     }
 }
